move unit tests to tests/unit
