+++ tests/unit/server/settings_test.py
