+++ tox.ini
@@ -20,7 +20,7 @@ commands =
     uv --version
     python --version
     pytest --version
-    pytest -n auto --disable-warnings
+    pytest -n auto --disable-warnings tests/unit
 
 [testenv:syft.jupyter]
 description = Jupyter Notebook with Editable Syft
